server/requirements.txt
Flask
picamera2
opencv-python
RPi.GPIO
numpy